+++ source/client/eterBase/tea.h
@@ -11,8 +11,8 @@ extern "C" {
 
 	#define TEA_KEY_LENGTH 16
 
-    int tea_encrypt(unsigned long *dest, const unsigned long *src, const unsigned long *key, int size);
-    int tea_decrypt(unsigned long *dest, const unsigned long *src, const unsigned long *key, int size);
+    int32_t tea_encrypt(uint32_t *dest, const uint32_t *src, const uint32_t *key, int32_t size);
+    int32_t tea_decrypt(uint32_t *dest, const uint32_t *src, const uint32_t *key, int32_t size);
 
 #ifdef __cplusplus
 };
